+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - New column for the main channels table: ``PERC_US``: the percentage of funds in the channel that belong to us.
+- New column ``TOTAL_SATS``: the total size of the channel in sats
 
 ## [3.3.1] 2024-06-29